Fix dimension ordering for warped 4D timeseries output
The warped volumes are extracted in [z,y,x] format (SimpleITK's natural
ordering from GetArrayFromImage), not [H,W,D]. When stacking and saving
as 4D NIfTI, we need to transpose from [z,y,x,T] to [T,z,y,x], which is
(3,0,1,2) not (3,2,0,1).

This fixes the dimension scrambling where spatial and temporal dimensions
were getting mixed up in the output.